- chunk.c patches from #1510
git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2033 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/chunk.h
@@ -67,7 +67,7 @@ LI_API buffer * chunkqueue_get_append_buffer(chunkqueue *c);
 LI_API buffer * chunkqueue_get_prepend_buffer(chunkqueue *c);
 LI_API chunk * chunkqueue_get_append_tempfile(chunkqueue *cq);
 LI_API int chunkqueue_steal_tempfile(chunkqueue *cq, chunk *in);
-LI_API int chunkqueue_steal_chunk(chunkqueue *cq, chunk *c);
+LI_API off_t chunkqueue_steal_chunk(chunkqueue *cq, chunk *c);
 LI_API off_t chunkqueue_steal_chunks_len(chunkqueue *cq, chunk *c, off_t max_len);
 LI_API off_t chunkqueue_steal_all_chunks(chunkqueue *cq, chunkqueue *in);
 LI_API off_t chunkqueue_skip(chunkqueue *cq, off_t skip);
